cleanup and refactor
